+++ constants/vo.go
@@ -1,0 +1,4 @@
+package constants
+
+type Severity string
+type Health string
