+++ validator/vo.go
@@ -4,8 +4,6 @@ type SysType string
 type ModelType string
 type ModelID string
 type MessageCode string
-type Severity string
-type Health string
 
 type ValidationResults map[ModelType]map[ModelID]*ValidationResult
